The Management System of College Student Status Based
The Management System of College Student Status Based 
on JavaWeb
Abstract
Computer technology has developed at high speed in the field of information science, and as time goes on, computer technology has became more and more mature and has been widely used by various industries. It brings great convenience to all fields and integrates into the bone marrow of these fields. In order to coordinate with all aspects of the education industry, realized automation in the education industry and reduced labor costs, this paper designed and developed a university-based student status management system based on JavaWeb.
This design described the research content, researched significance, related technologies used in the development of university-based student management system based on JavaWeb, the specific design framework and implementation processed of the system, and the overall debugged of the system. The main purpose was to let the reader understand the main structure and framework of this system, and this system is playing its due role. In this paper, detailed design and optimization of user management, teacher management, and performance management of the system based on Java Web to achieved university student status management system were carried out. The system flow, implementation diagram and related code pairs are described in detail. In the database, this design used the MySQL database, combined with html and ajax, used Java language for system development, can achieved the basic functions of the University-based student management system based on JavaWeb.
The developed university web-based management system based on Java Web is friendly, easy to operate, and meets the needs of users in all aspects. It brings great convenience to the management of the education industry.

Keywords: School Status ;Performance Management;Student Status Management System